% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clearColnames.R
\name{clearColnames}
\alias{clearColnames}
\title{Służy do uporządkowania nazw kolumn}
\usage{
clearColnames(x, removePolishSings = T)
}
\arguments{
\item{x}{vector numeryczny (numeric vector)}
}
\value{
zwraca vector po usunięciu wszystkich polskich znaków (zamienieniu ich na odpowiednik ą -> a)
}
\description{
Służy do uporządkowania nazw kolumn
}
