% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dodajSeparatorTYS.R
\name{dodajSeparatorTYS}
\alias{dodajSeparatorTYS}
\title{Formatuje jak PLN}
\usage{
dodajSeparatorTYS(x, suffix = "", accuracy = 2, scale = 1,
  prefix = "", big.mark = " ", largest_with_cents = 1000 * 1000,
  decimal.mark = ",", negative_parens = FALSE, ...)
}
\arguments{
\item{suffix}{co ma byc po kwocie np. znak 'dolara' domyslnie ' zł'}
}
\value{
zwraca character vector dla kwot
}
\description{
Formatuje jak PLN
}
