% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drWhyPallete.R
\name{drWhyPallete}
\alias{drWhyPallete}
\title{drWhyPallete}
\usage{
drWhyPallete(N)
}
\arguments{
\item{x}{int od 1 do n (9)}
}
\value{
zwraca werktor kolorów z pallety drWhy
}
\description{
drWhyPallete to paleta kolorow wykorzystana w https://github.com/ModelOriented/DrWhy
W przypadku gdy uzytkownik wybierze wiecej niż N > 9, dodawane sa z kolory z rainbow pallete.
}
