% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epochToYMDHMS.R
\name{epochToYMDHMS}
\alias{epochToYMDHMS}
\title{Convert epoch to YMD HMS date format}
\usage{
epochToYMDHMS(x)
}
\arguments{
\item{x}{vector numeryczny (numeric vector)}
}
\value{
zwraca vector po usunieciu wszystkich polskich znakow (zamienieniu ich na odpowiednik a z sladczkiem -> a)
}
\description{
Convert epoch to YMD HMS date format
}
