% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatujKwote.R
\name{formatujKwote}
\alias{formatujKwote}
\title{Służy do formatowania kwot przy prezentacji na np. wykresach i danych zawsze zwraca character}
\usage{
formatujKwote(x, miesjcaPoPrzecinku = 1, prefix = "", suffix = "zł")
}
\arguments{
\item{x}{vector numeryczny (numeric vector)}

\item{prefix}{co ma być przed kwotą np. znak '$' domyślnie nic ''}

\item{suffix}{co ma być po kwocie np. znak '$' domyślnie nic 'zł'}

\item{miejscaPoPrzecinku}{ilość miejsc po przecinku, które mają być prezentowane domyślnie 1 (jedno)}
}
\value{
zwraca characet vector dla kwot
}
\description{
Służy do formatowania kwot przy prezentacji na np. wykresach i danych zawsze zwraca character
}
