% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggPlotMissing.R
\name{ggPlotMissings}
\alias{ggPlotMissings}
\title{ggPlotMissings}
\usage{
ggPlotMissings(data, title = NULL, ggtheme = ggplot2::theme_classic())
}
\arguments{
\item{data}{object class c("tbl_df", "tbl","data.frame")}

\item{title}{tytul wykresu}

\item{ggtheme}{funkcja theme z pakietu ggplot2}
}
\value{
zwraca wykres pokazujacy % ilosci NA w danych
}
\description{
Funkcja pomocnicza to pokazywania ile jest brakow danych w poszczegolnych kolumnach
}
