% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measureTime.R
\name{measureTime}
\alias{measureTime}
\title{Słuzy do formatowania kwot przy prezentacji na np. wykresach i danych zawsze zwraca character}
\usage{
measureTime(start = Sys.time(), end = NULL)
}
\arguments{
\item{x}{vector numeryczny (numeric vector)}

\item{miejscaPoPrzecinku}{ilosc miejsc po przecinku, ktore maja byc prezentowane domyclnie 1 (jedno)}

\item{prefix}{co ma byc przed kwota np. znak '$' domyslnie nic ''}

\item{suffix}{co ma byc po kwocie np. znak '$' domyslnie nic 'zł'}
}
\value{
zwraca characet vector dla kwot
}
\description{
Słuzy do formatowania kwot przy prezentacji na np. wykresach i danych zawsze zwraca character
}
