% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtsToDataFrame.R
\name{mtsToDataFrame}
\alias{mtsToDataFrame}
\title{Konwersja ts lub mts do data.frame}
\usage{
mtsToDataFrame(x)
}
\arguments{
\item{x}{mts, or ts object}
}
\value{
data.frame of converted ts, mts object
}
\description{
Konwersja ts lub mts do data.frame
}
