% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLN.R
\name{PLN}
\alias{PLN}
\title{Formatuje jak PLN}
\usage{
PLN(x, suffix = " zł", accuracy = 2, scale = 1, prefix = "",
  big.mark = " ", largest_with_cents = 1000 * 1000,
  decimal_mark = ",", negative_parens = FALSE, trim = TRUE, ...)
}
\arguments{
\item{suffix}{co ma byc po kwocie np. znak 'dolara' domyslnie ' zł'}
}
\value{
zwraca character vector dla kwot
}
\description{
Formatuje jak PLN
}
