% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotColours.R
\name{plotColors}
\alias{plotColors}
\title{plotColors}
\usage{
plotColors(palette_var)
}
\arguments{
\item{pallete_var}{character vector z paletami}
}
\value{
zwraca object klasy gg, ggplot
}
\description{
plotColors to funkcja, która obrazuje kolory w palecie, która jest dostarczona
}
