% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFilesWithCustomFUN.R
\name{readFilesWithCustomFUN}
\alias{readFilesWithCustomFUN}
\title{Funkcja służy jako generyk to wczytywania wielu plików za pomocą różnych funkcji, można tam wstawić csv, xlsx itp}
\usage{
readFilesWithCustomFUN(paths, FUN, ...)
}
\arguments{
\item{paths}{wektor ścieżek do plików}

\item{FUN}{niestandardowa funkcja np. readxl::read_excel, czy read.csv}

\item{...}{dodatkowe parametry}
}
\value{
zwraca listę z wczytanymi plikami.
}
\description{
Funkcja służy jako generyk to wczytywania wielu plików za pomocą różnych funkcji, można tam wstawić csv, xlsx itp
}
