% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usuwaniePolskichZnakow.R
\name{usuwaniePolskichZnakow}
\alias{usuwaniePolskichZnakow}
\title{Służy do zamiany polskich liter na ich odpowiedniki bez śladczków}
\usage{
usuwaniePolskichZnakow(x, perl = F)
}
\arguments{
\item{x}{vector numeryczny (numeric vector)}
}
\value{
zwraca vector po usunięciu wszystkich polskich znaków (zamienieniu ich na odpowiednik ą -> a)
}
\description{
Służy do zamiany polskich liter na ich odpowiedniki bez śladczków
}
