% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatujKwote.R
\name{formatujKwote}
\alias{formatujKwote}
\title{Sluzy do formatowania kwot przy prezentacji na np. wykresach i danych zawsze zwraca character}
\usage{
formatujKwote(x, miejscaPoPrzecinku = 1, prefix = "", suffix = "zł")
}
\arguments{
\item{x}{vector numeryczny (numeric vector)}

\item{miejscaPoPrzecinku}{ilosc miejsc po przecinku, ktore maja byc prezentowane domyslnie 1 (jedno)}

\item{prefix}{co ma byc przed kwota np. string "zl" domyslnie nic ''}

\item{suffix}{co ma byc po kwocie np. znak 'dolara' domyslnie 'zł'}
}
\value{
zwraca character vector dla kwot
}
\description{
Sluzy do formatowania kwot przy prezentacji na np. wykresach i danych zawsze zwraca character
}
